Backend: Controller, Service y Client de Users ahora no ignoran errores y termine de implementar interface, cambios en los dto y nombres de algunas variables. En el servicio de Courses tambien agregue un poco de errores pero falta.

+++ backend/dto/dto_user_course_min.go
@@ -1,11 +1,11 @@
 package dto
 
 type UserCourseMinDto struct {
-	IdUser   int    `json:"id_user"`
-	IdCourse int    `json:"id_course"`
-	Rating   int    `json:"rating"`
-	Comment  string `json:"comment"`
-	IsActive bool   `json:"is_active"`
+	IdUser   int     `json:"id_user"`
+	IdCourse int     `json:"id_course"`
+	Rating   float64 `json:"rating"`
+	Comment  string  `json:"comment"`
+	IsActive bool    `json:"is_active"`
 }
 
 type UserCoursesMinDto []UserCourseMinDto
